Update README.md
A little concerned, it states that the commits need to be completed by 5th February however the assignment spec' on the VLE states the following:

 a commit timestamp needs to be within 2 weeks of the allocated Week (i.e. week
1 needs completing before the end of Week 3 etc)

This is under "Requirement 7". There exists no Requirement 6 which also caused some concern however this may be due to the previous paragraph which caused confusion with the numbers.